pyproject.toml
[tool.poetry]
name = "emronfhir"
version = "0.1.0"
description = ""
authors = ["Zbynek Vyhlas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.6"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
python-decouple = "^3.8"
djangorestframework = "^3.14.0"
pillow = "^10.0.0" # enabling django work with images
dj-rest-auth = "^4.0.1"
djangorestframework-simplejwt = "^5.3.0"
django-sslserver = "^0.22"
django-allauth = "0.54.0"
celery = "^5.3.4"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
django-import-export = "^3.3.1"
colorlog = "^6.7.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
watchdog = {extras = ["watchmedo"], version = "^3.0.0"}
faker = "^19.12.1"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
